Download CoralTemp data

Function to download and save NetCDF files for CoralTemp download_CoralTemp() is a function to download NOAA CRW data

If the number of cores is set to >1, the function uses mclapply to parallel download datasets, else 1 = single downloads.

Notes: The NOAA Coral Reef Watch (CRW) daily global 5km Sea Surface Temperature (SST) product, also known as CoralTemp, shows the nighttime ocean temperature measured at the surface. The CoralTemp SST data product was developed from two, related reanalysis (reprocessed) SST products and a near real-time SST product. Spanning January 1, 1985 to the present, the CoralTemp SST is one of the best and most internally consistent daily global 5km SST products available.

Usage

download_CoralTemp(
  url = "https://www.ncei.noaa.gov/thredds-ocean/catalog/crw/5km/v3.1/nc/v1.0/daily",
  start_date,
  end_date,
  dest_dir,
  variable = "hs",
  mc.cores = 1
)

Arguments

url

one of the NOAA thredds url

start_date

end year

end_date

end year

dest_dir

save file location

variable

one of 'sst', 'ssta', 'hs', 'dhw'

mc.cores

set the number of cores

Value

downloaded nc files to specified location

Examples

if (FALSE) { # \dontrun{
download_CoralTemp(1990, 2020, "/users/era5/", variable= "ssta", cores=10)
} # }